test_dev_MT/pyproject.toml
[tool.poetry]
name = "test-dev-mt"
version = "0.1.0"
description = ""
authors = ["Henrik Anfinsen <henrik@solgt.no>"]
readme = "README.md"
packages = [{include = "test_dev_mt"}]

[tool.poetry.dependencies]
python = ">=3.9.0,<3.11"
pymongo = "^4.4.1"
matplotlib = "3.7.2"
geopandas = "0.13.2"
"solgt.db" = {path = "../../py/lib/db", develop = true}
"solgt.priceindex" = {path = "../../py/lib/priceindex", develop = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
